chg: Make WaitForClientState part of the Client interface
